[refactor] Renamed aux package to resolve Windows issue
